+++ src/Logitar.Portal.Core/Configurations/ConfigurationAlreadyInitializedException.cs
@@ -1,0 +1,8 @@
+﻿namespace Logitar.Portal.Core.Configurations;
+
+public class ConfigurationAlreadyInitializedException : Exception
+{
+  public ConfigurationAlreadyInitializedException() : base("The configuration has already been initialized.")
+  {
+  }
+}