+++ src/Logitar.Portal.Core/Configurations/Queries/ProjectToConfigurationOutput.cs
@@ -1,0 +1,5 @@
+﻿using MediatR;
+
+namespace Logitar.Portal.Core.Configurations.Queries;
+
+public record ProjectToConfigurationOutput(ConfigurationAggregate Configuration) : IRequest<Configuration>;