+++ src/Logitar.Portal.Core/Configurations/Validators/ConfigurationInitializedValidator.cs
@@ -1,19 +1,10 @@
-﻿using FluentValidation;
-using Logitar.Portal.Core.Configurations.Events;
-using Logitar.Portal.Core.Realms.Validators;
+﻿using Logitar.Portal.Core.Configurations.Events;
 
 namespace Logitar.Portal.Core.Configurations.Validators;
 
-internal class ConfigurationInitializedValidator : AbstractValidator<ConfigurationInitialized>
+internal class ConfigurationInitializedValidator : ConfigurationSavedValidator<ConfigurationInitialized>
 {
-  public ConfigurationInitializedValidator()
+  public ConfigurationInitializedValidator() : base()
   {
-    RuleFor(x => x.DefaultLocale).Locale();
-
-    RuleFor(x => x.UsernameSettings).SetValidator(new ReadOnlyUsernameSettingsValidator());
-
-    RuleFor(x => x.PasswordSettings).SetValidator(new ReadOnlyPasswordSettingsValidator());
-
-    RuleFor(x => x.LoggingSettings).SetValidator(new ReadOnlyLoggingSettingsValidator());
   }
 }