+++ src/Logitar.Portal.EntityFrameworkCore.PostgreSQL/Commands/MigrateDatabase.cs
@@ -1,5 +1,0 @@
-﻿using MediatR;
-
-namespace Logitar.Portal.EntityFrameworkCore.PostgreSQL.Commands;
-
-public record MigrateDatabase : IRequest;